Progress on #3.
Can now retrieve the set of receivers from a call graph node. The set is a set of InstanceKeys. Using these, we can look up their state in the associated factoids. We probably can also get the receivers of those instances when they were created because InstanceKeys are simply nodes themselves but I haven't tested that just yet.