Caching for #12.
Not exactly sure if this is helpful just yet because we don't yet know
who the receiver is for the terminal operation. However, perhaps we
should fill out the entire table and then go looking for terminal
operations? But, really, we are only interested in the state at such
operations. Hm.