Fixed #38.
Really just a workaround, though. The solution is to use generic type
information from the source code to filter out points-to set members
that could never be assignable to the pointer. The matching is
potentially inaccurate, however, since we don't have column information
from the SSA. The matching is currently only done on the line level.
Thus, having multiple terminal operations on the same line could be
problematic but (thankfully) not likely.